Fix rem / em css, rename scss files as module.scss
