+++ src/components/Layout/Layout.module.scss
@@ -1,5 +1,5 @@
 .layout {
   margin: 0 auto;
   font-family: roboto;
-  max-width: 100rem;
+  max-width: 100%;
 }
