+++ src/components/tasks/section/section.module.scss
@@ -3,12 +3,13 @@
   font-weight: bold;
   text-align: center;
   color: #041484;
-  margin-bottom: 2rem;
+  margin-bottom: 1em;
 }
 
 .section {
   height: 100%;
-  min-width: 17rem;
+  min-width: 17em;
+  font-size: 1rem;
 }
 
 .cardContainer {
