+++ src/styles/index.scss
@@ -4,12 +4,14 @@ body {
 
 html {
   font-size: 15px;
-  margin: 0 5rem;
+  margin: 0 5em;
 }
+
 .container {
   display: flex;
   flex-direction: row;
   justify-content: space-evenly;
   align-items: flex-start;
   flex-wrap: wrap;
+  font-size: 1rem;
 }
